requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain==0.3.4
langchain-core==0.3.12
langchain-community==0.3.3
faiss-cpu==1.9.0
langchain_google_genai==2.0.1
sentence_transformers==3.2.1